Move pattern module to provider::pattern
