+++ components/datetime/src/format/neo.rs
@@ -14,9 +14,9 @@ use crate::input;
 use crate::input::ExtractedInput;
 use crate::neo_pattern::{DateTimePattern, DateTimePatternBorrowed};
 use crate::neo_skeleton::NeoDateTimeSkeleton;
-use crate::pattern::PatternItem;
 use crate::provider::neo::*;
 use crate::provider::time_zones::tz;
+use crate::provider::pattern::PatternItem;
 use crate::scaffold::*;
 use crate::scaffold::{
     AllInputMarkers, DateInputMarkers, DateTimeMarkers, GetField, IsInCalendar, NeoNeverMarker,
