+++ components/datetime/src/provider/pattern/item/mod.rs
@@ -5,14 +5,14 @@
 mod generic;
 mod ule;
 
+use super::PatternError;
 use crate::fields::{Field, FieldLength, FieldSymbol};
-use crate::pattern::PatternError;
 use core::convert::TryFrom;
 pub use generic::GenericPatternItem;
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 #[cfg_attr(feature = "datagen", derive(serde::Serialize, databake::Bake))]
-#[cfg_attr(feature = "datagen", databake(path = icu_datetime::pattern::item))]
+#[cfg_attr(feature = "datagen", databake(path = icu_datetime::provider::pattern::item))]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize))]
 #[allow(clippy::exhaustive_enums)] // this type is stable
 pub enum PatternItem {
