+++ components/datetime/src/provider/pattern/reference/pattern.rs
@@ -2,7 +2,7 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use crate::pattern::runtime;
+use super::super::runtime;
 
 use super::{
     super::{PatternError, PatternItem, TimeGranularity},
