+++ components/datetime/src/provider/pattern/runtime/plural.rs
@@ -2,10 +2,8 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use crate::{
-    fields::{Field, FieldSymbol, Week},
-    pattern::{runtime::Pattern, PatternError, PatternItem},
-};
+use super::super::{runtime::Pattern, PatternError, PatternItem};
+use crate::fields::{Field, FieldSymbol, Week};
 use either::Either;
 use icu_plurals::PluralCategory;
 use icu_provider::prelude::*;
@@ -13,7 +11,7 @@ use icu_provider::prelude::*;
 /// A collection of plural variants of a pattern.
 #[derive(Debug, PartialEq, Clone, yoke::Yokeable, zerofrom::ZeroFrom)]
 #[cfg_attr(feature = "datagen", derive(serde::Serialize, databake::Bake))]
-#[cfg_attr(feature = "datagen", databake(path = icu_datetime::pattern::runtime))]
+#[cfg_attr(feature = "datagen", databake(path = icu_datetime::provider::pattern::runtime))]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize))]
 #[allow(clippy::exhaustive_structs)] // part of data struct
 pub struct PluralPattern<'data> {
@@ -122,7 +120,7 @@ impl<'data> PluralPattern<'data> {
 #[allow(clippy::large_enum_variant)]
 #[allow(clippy::exhaustive_enums)] // this type is stable
 #[cfg_attr(feature = "datagen", derive(databake::Bake))]
-#[cfg_attr(feature = "datagen", databake(path = icu_datetime::pattern::runtime))]
+#[cfg_attr(feature = "datagen", databake(path = icu_datetime::provider::pattern::runtime))]
 pub enum PatternPlurals<'data> {
     /// A collection of pattern variants for when plurals differ.
     MultipleVariants(PluralPattern<'data>),
