+++ provider/source/src/datetime/patterns.rs
@@ -3,8 +3,8 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 use crate::cldr_serde;
-use icu::datetime::pattern;
-use icu::datetime::pattern::CoarseHourCycle;
+use icu::datetime::provider::pattern;
+use icu::datetime::provider::pattern::CoarseHourCycle;
 use icu::datetime::provider::calendar::*;
 
 impl From<&cldr_serde::ca::LengthPatterns> for patterns::LengthPatternsV1<'_> {
